move ModuleLaunchPad to _comps dir
